+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New checks:
 * ``consider-using-background``
 * ``examples-outside-scenario-outline``
 * ``duplicated-tag``
+* ``duplicated-scenario-name``
 
 Other changes:
 * ``unparseable-file`` is now issued for all nodes that offend the ``gherkin`` parser together with more error information provided by it